(PDB-1902) Create a defaulted config service
Currently configuration is retrieved from the TK config service,
defaulted and consumed by other componentes via shared-globals. This
commit creates a new DefaultedConfig service that has a similar API as
the config service, but applies our validation and defaulting logic
first. This will allow a large amount of shared-globals to be removed
and instead just use the defaulted values from this new service.